+++ example/templates/currentTime.ts
@@ -1,28 +1,11 @@
 import { Type } from '@fastify/type-provider-typebox';
-import type { Template } from 'nettletea';
+import type { Template } from '../../src/types.js';
 
 export interface Data {
 	currentTime: string;
 }
 
 export default {
-	name: 'Current Time',
-	fn: ({ currentTime }: Data) => {
-		const text = `Now is ${new Date(Date.parse(currentTime)).toLocaleString('de')}!`;
-		return {
-			blocks: [
-				{
-					text: {
-						type: 'mrkdwn',
-						text: text,
-					},
-					type: 'section',
-				},
-			],
-			text,
-		};
-	},
-
 	examples: {
 		default: {
 			currentTime: new Date().toISOString(),
@@ -31,6 +14,22 @@ export default {
 			currentTime: new Date(2024, 1, 1).toISOString(),
 		},
 	},
+	fn: ({ currentTime }: Data) => {
+		const text = `Now is ${new Date(Date.parse(currentTime)).toLocaleString('de')}!`;
+		return {
+			blocks: [
+				{
+					text: {
+						text: text,
+						type: 'mrkdwn',
+					},
+					type: 'section',
+				},
+			],
+			text,
+		};
+	},
+	name: 'Current Time',
 	schema: Type.Object({
 		currentTime: Type.String({ format: 'date-time' }),
 	}),