+++ example/templates/index.ts
@@ -1,4 +1,0 @@
-export const all = {
-	currentTime: (await import('./currentTime')).default,
-	helloWorld: (await import('./helloworld')).default,
-};